Entry: - change workTime and overTime variable names to specify unit of measure (seconds) and type to Int
EntryMO:
- change overTime and workTime type to Int64

+++ ClockIn/Model/EntryMO+CoreDataProperties.swift
@@ -17,9 +17,9 @@ extension EntryMO {
     }
 
     @NSManaged public var finishDate: Date
-    @NSManaged public var overTime: Double
+    @NSManaged public var overTime: Int64
     @NSManaged public var startDate: Date
-    @NSManaged public var workTime: Double
+    @NSManaged public var workTime: Int64
     @NSManaged public var id: UUID
 
 }
